cleanup

+++ src/sedec2/arib/b10/TableFactory.java
@@ -30,99 +30,99 @@ import sedec2.base.Table;
 
 public class TableFactory {
     /** @note PAT */
-    public final static int PROGRAM_ASSOCIATION_TABLE = 0x00;
+    public final static byte PROGRAM_ASSOCIATION_TABLE = (byte) 0x00;
     
     /** @note CAT */
-    public final static int CONDITIONAL_ACCESS_TABLE = 0x01;
+    public final static byte CONDITIONAL_ACCESS_TABLE = (byte) 0x01;
     
     /** @note PMT */
-    public final static int PROGRAM_MAP_TABLE = 0x02;
+    public final static byte PROGRAM_MAP_TABLE = (byte) 0x02;
     
     /** @note NIT for actual */
-    public final static int ACTUAL_NETWORK_INFORMATION_TABLE = 0x40;
+    public final static byte ACTUAL_NETWORK_INFORMATION_TABLE = (byte) 0x40;
     
     /** @note NIT for other */
-    public final static int OTHER_NETWORK_INFORMATION_TABLE = 0x41;
+    public final static byte OTHER_NETWORK_INFORMATION_TABLE = (byte) 0x41;
     
     /** @note SDT for actual */
-    public final static int ACTUAL_SERVICE_DESCRIPTION_TABLE = 0x42;
+    public final static byte ACTUAL_SERVICE_DESCRIPTION_TABLE = (byte) 0x42;
     
     /** @note SDT for other */
-    public final static int OTHER_SERVICE_DESCRIPTION_TABLE = 0x46;
+    public final static byte OTHER_SERVICE_DESCRIPTION_TABLE = (byte) 0x46;
     
     /** @note BAT */
-    public final static int BOUQUET_ASSOCIATION_TABLE = 0x4a;
+    public final static byte BOUQUET_ASSOCIATION_TABLE = (byte) 0x4a;
     
     /** @note EIT for actual present */
-    public final static int ACTUAL_TS_PRESENT_EVENT_INFORMATION_TABLE = 0x4e;
+    public final static byte ACTUAL_TS_PRESENT_EVENT_INFORMATION_TABLE = (byte) 0x4e;
     
     /** @note EIT for other */
-    public final static int OTHER_TS_PRESENT_EVENT_INFORMATION_TABLE = 0x4f;
+    public final static byte OTHER_TS_PRESENT_EVENT_INFORMATION_TABLE = (byte) 0x4f;
     
     /** @note TDT */
-    public final static int TIME_DATE_TABLE = 0x70;
+    public final static byte TIME_DATE_TABLE = (byte) 0x70;
     
     /** @note RST */
-    public final static int RUNNING_STATUS_TABLE = 0x71;
+    public final static byte RUNNING_STATUS_TABLE = (byte) 0x71;
     
     /** @note ST */
-    public final static int STUFFING_TABLE = 0x72;
+    public final static byte STUFFING_TABLE = (byte) 0x72;
     
     /** @note TOT */
-    public final static int TIME_OFFSET_TABLE = 0x73;
+    public final static byte TIME_OFFSET_TABLE = (byte) 0x73;
     
     /** @note AIT */
-    public final static int APPLICATION_INFORMATION_TABLE = 0x74;
+    public final static byte APPLICATION_INFORMATION_TABLE = (byte) 0x74;
     
     /** @note DIT */
-    public final static int DISCONTINUITY_INFORMATION_TABLE = 0x7e;
+    public final static byte DISCONTINUITY_INFORMATION_TABLE = (byte) 0x7e;
     
     /** @note SIT */
-    public final static int SELECTION_INFORMATION_TABLE = 0x7f;
+    public final static byte SELECTION_INFORMATION_TABLE = (byte) 0x7f;
     
     /** @note ECM */
-    public final static int ENTITLEMENT_CONTROL_MESSAGE_1 = 0x82;
-    public final static int ENTITLEMENT_CONTROL_MESSAGE_2 = 0x83;
+    public final static byte ENTITLEMENT_CONTROL_MESSAGE_1 = (byte) 0x82;
+    public final static byte ENTITLEMENT_CONTROL_MESSAGE_2 = (byte) 0x83;
 
     /** @note EMM */
-    public final static int ENTITLEMENT_MANAGEMENT_MESSAGE_1 = 0x84;
-    public final static int ENTITLEMENT_MANAGEMENT_MESSAGE_2 = 0x85;
+    public final static byte ENTITLEMENT_MANAGEMENT_MESSAGE_1 = (byte) 0x84;
+    public final static byte ENTITLEMENT_MANAGEMENT_MESSAGE_2 = (byte) 0x85;
 
     /** @note DCM */
-    public final static int DOWNLOAD_CONTROL_TABLE = 0xc0;
+    public final static byte DOWNLOAD_CONTROL_TABLE = (byte) 0xc0;
     
     /** @note DT */
-    public final static int DOWNLOAD_TABLE = 0xc1;
+    public final static byte DOWNLOAD_TABLE = (byte) 0xc1;
     
     /** @note PCAT */
-    public final static int PARTIAL_CONTENT_ANNOUNCEMENT_TABLE = 0xc2;
+    public final static byte PARTIAL_CONTENT_ANNOUNCEMENT_TABLE = (byte) 0xc2;
     
     /** @note BIT */
-    public final static int NETWORK_BOARD_INFORMATION_TABLE_1 = 0xc5;
-    public final static int NETWORK_BOARD_INFORMATION_TABLE_2 = 0xc6;
+    public final static byte NETWORK_BOARD_INFORMATION_TABLE_1 = (byte) 0xc5;
+    public final static byte NETWORK_BOARD_INFORMATION_TABLE_2 = (byte) 0xc6;
 
     /** @note LDT */
-    public final static int LINKED_DESCRIPTION_TABLE = 0xc7;
+    public final static byte LINKED_DESCRIPTION_TABLE = (byte) 0xc7;
 
     /** @note LIT */
-    public final static int LOCAL_EVENT_INFORMATION_TABLE = 0xd0;
+    public final static byte LOCAL_EVENT_INFORMATION_TABLE = (byte) 0xd0;
     
     /** @note ERT */
-    public final static int EVENT_RELATION_TABLE = 0xd1;
+    public final static byte EVENT_RELATION_TABLE = (byte) 0xd1;
     
     /** @note ITT */
-    public final static int INDEX_TRANSMISSION_TABLE = 0xd2;
+    public final static byte INDEX_TRANSMISSION_TABLE = (byte) 0xd2;
     
     /** @note SDTT */
-    public final static int SOFTWARE_DOWNLOAD_TRIGGER_TABLE = 0xc3;
+    public final static byte SOFTWARE_DOWNLOAD_TRIGGER_TABLE = (byte) 0xc3;
     
     /** @note BIT */
-    public final static int BROADCASTER_INFORMATION_TABLE = 0xc4;
+    public final static byte BROADCASTER_INFORMATION_TABLE = (byte) 0xc4;
     
-    public final static int UNKNOWN_TABLE = 0xff;
+    public final static byte UNKNOWN_TABLE = (byte) 0xff;
     
     public static Table createTable(byte[] buffer) {
-        int table_id = (buffer[0] & 0xff);
+        byte table_id = (byte)(buffer[0] & 0xff);
         
         Table section = null;
